npm ut refactor
