+++ npm/testpolicies/allow-all-to-app-frontend.yaml
@@ -1,0 +1,13 @@
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  name: ALLOW-all-TO-app:frontend-FROM-all-namespaces-policy
+  namespace: "testnamespace"
+spec:
+  podSelector:
+    matchLabels:
+      app: "frontend"
+  ingress:
+    - {}
+  policyTypes:
+    - Ingress
