+++ npm/testpolicies/allow-backend-to-frontend.yaml
@@ -1,0 +1,16 @@
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  name: deny-all-policy
+  namespace: "testnamespace"
+spec:
+  podSelector:
+    matchLabels:
+      app: "backend"
+  ingress:
+    - from:
+        - podSelector:
+            matchLabels:
+              app: frontend
+  policyTypes:
+    - Ingress
