+++ npm/testpolicies/allow-ns-dev-and-app-backend-to-app-frontend.yaml
@@ -1,0 +1,19 @@
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  name: deny-all-policy
+  namespace: "testnamespace"
+spec:
+  policyTypes:
+    - Ingress
+  podSelector:
+    matchLabels:
+      app: "frontend"
+  ingress:
+    - from:
+        - podSelector:
+            matchLabels:
+              app: backend
+          namespaceSelector:
+            matchLabels:
+              ns: dev
