+++ npm/testpolicies/deny-all-from-app-backend.yaml
@@ -1,0 +1,11 @@
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  name: deny-all-policy
+  namespace: "testnamespace"
+spec:
+  policyTypes:
+    - Egress
+  podSelector:
+    matchLabels:
+      app: "backend"
