+++ npm/testpolicies/deny-all-from-ns-unsafe.yaml
@@ -1,0 +1,10 @@
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  name: deny-all-policy
+  namespace: "unsafe"
+spec:
+  policyTypes:
+    - Egress
+  podSelector: {}
+  egress: []
