+++ npm/testpolicies/deny-all-policy.yaml
@@ -1,0 +1,10 @@
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  name: deny-all-policy
+  namespace: "testnamespace"
+spec:
+  policyTypes:
+    - Ingress
+  podSelector: {}
+  ingress: []
