+++ npm/testpolicies/deny-all-to-app-frontend.yaml
@@ -1,0 +1,11 @@
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  name: deny-all-policy
+  namespace: "testnamespace"
+spec:
+  podSelector:
+    matchLabels:
+      app: "frontend"
+  policyTypes:
+    - Ingress
